finalized credit/debit use case inside CardReaderNotifications, Session, and TransactionRecord classes

+++ com.thelocalmarketplace.software/src/com/thelocalmarketplace/software/TransactionRecord.java
@@ -26,7 +26,7 @@ public class TransactionRecord {
     // default variables
     private static long totalPrice = 0;
     private static ArrayList<BarcodedProduct> bill;
-    private static int holdNumber = 0;
+    private static long holdNumber = 0;
     private static int paymentType = NONE;
     private static BigDecimal amountOwed = new BigDecimal("0");
 
@@ -53,7 +53,7 @@ public class TransactionRecord {
     }
 
     // methods that modify hold number
-    public void updateHoldNumber(int number) {
+    public void updateHoldNumber(long number) {
         holdNumber = number;
     }
 
@@ -113,7 +113,7 @@ public class TransactionRecord {
         return bill.get(index);
     }
 
-    public int getHoldNumber() {
+    public long getHoldNumber() {
         return holdNumber;
     }
